requirements.txt
pandas>=2.0
requests>=2.28
python-dateutil>=2.8
scikit-learn>=1.3
spacy>=3.7
vaderSentiment>=3.3.2
python-docx>=1.1.0
